requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
imgkit==1.2.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
playwright==1.45.0
pluggy==1.5.0
priority==2.0.0
pycparser==2.22
pyee==11.1.0
PySocks==1.7.1
pytest==8.2.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dotenv==1.0.1
python-slugify==8.0.4
Quart==0.19.6
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4